+++ source/standard/expressions/ThymeleafExpressionLanguage.js
@@ -247,8 +247,12 @@ export default new Grammar('Thymeleaf Expression Language',
 
 	/**
 	 * String literal, characters surrounded by `'` (single quotes).
+	 * 
+	 * This is trying to emulate negative lookbehind so that escaped quotes don't
+	 * get counted as string terminators, but JavaScript only got that feature in
+	 * ES2018, so if I used it it'd leave too many JS engines without support.
 	 */
-	new ThymeleafRule('StringLiteral', /'.*?'/, result => () => result.slice(1, -1)),
+	new ThymeleafRule('StringLiteral', /'.*?(?!\\').'/, result => () => result.slice(1, -1).replace(/\\/g, '')),
 
 	/**
 	 * A number.